pyproject.toml
[project]
name = "awtrix"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "paho-mqtt==2.1.0",
    "python-dotenv==1.0.1",
    "requests>=2.32.3",
    "wikiquote>=0.1.17",
]
